feat(k): add dbg and dashboard subcommands to kubectl wrapper
`dbg` is a wrapper around the existing `debug` subcommand, which
allows to interactively debug a container by adding an ephemeral
container to an already running pod. This is extended by the
functionality to automatically select the target container in
case that the pod only contains a single container and offers
a fuzzy selection of the target container using fzf in case there
are multiple containers running in the pod. A `busybox` image is used
for the debugging container.

`dashboard` is a custom subcommand that launches the [Kubernetes
Dashboard](https://github.com/kubernetes/dashboard) in case that a
kubernetes-dashboard service is running in a namespace with the same
name. It will port-forward the running services port 443 to 8443 on
localhost and open the resulting url in the default application. If it
doesn't find a corresponding service it will print the link to the
installation instructions.

The `get-pod-names` subcommand is removed in favor of using `--output name`.

The `get-pod-containers` subcommand is renamed to just `pod-containers`
in order to shorten the name and improve auto complete functionality
after the first few characters have been typed.